removed search bar in home and added title/search bar in trending and fixed misspelling for trending
